requirements.txt
Django
psycopg2
djangorestframework
markdown
django-filter
flake8
